requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
bandit==1.7.6
black==23.12.1
boltons==21.0.0
bracex==2.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
click-option-group==0.5.6
colorama==0.4.6
coverage==7.4.0
cryptography==41.0.4
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.5
django-allauth==0.57.0
django-cors-headers==4.2.0
django-extensions==3.2.3
django-phonenumber-field==7.2.0
django-user-agents==0.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.2.0
executing==2.0.1
face==22.0.0
frozenlist==1.4.1
geoip2==4.1.0
gitdb==4.0.11
GitPython==3.1.41
glom==22.1.0
idna==3.4
iniconfig==2.0.0
ipython==8.20.0
isort==5.13.2
jedi==0.19.1
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
matplotlib-inline==0.1.6
maxminddb==2.5.2
mdurl==0.1.2
multidict==6.0.4
mypy==1.8.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
parso==0.8.3
pathspec==0.12.1
pbr==6.0.0
peewee==3.17.0
pexpect==4.9.0
phonenumbers==8.13.23
Pillow==10.0.0
platformdirs==4.1.0
pluggy==1.3.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pyotp==2.9.0
pytest==7.4.4
pytest-cov==4.1.0
pytest-django==4.7.0
pytest-mock==3.11.1
python-lsp-jsonrpc==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
requests-mock==1.11.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.17.1
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
semgrep==1.34.1
six==1.16.0
smmap==5.0.1
sqlparse==0.4.4
stack-data==0.6.3
stevedore==5.1.0
tomli==2.0.1
traitlets==5.14.1
typing_extensions==4.9.0
ua-parser==0.18.0
ujson==5.9.0
urllib3==1.26.18
user-agents==2.2.0
wcmatch==8.5
wcwidth==0.2.13
whitenoise==6.6.0
yarl==1.9.4
